refactor: HTTP 핸들러 패키지 구조 변경
