+++ tomcat/src/main/java/org/apache/coyote/http11/handle/handler/StaticResourceHandler.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package org.apache.coyote.http11.handle.handler;
 
 import java.io.File;
 import java.io.IOException;
